change enum Board_type to uint32_t

+++ include/iit/ecat/advr/ft6_esc.h
@@ -203,7 +203,7 @@ public:
         return SDOs;
     }
     virtual void init_SDOs ( void );
-    virtual uint16_t get_ESC_type() {
+    virtual uint32_t get_ESC_type() {
         return FT6;
     }
