+++ include/iit/ecat/advr/power_board.h
@@ -156,7 +156,7 @@ public:
     virtual const objd_t * get_SDOs() {
         return SDOs;
     }
-    virtual uint16_t get_ESC_type() {
+    virtual uint32_t get_ESC_type() {
         return POW_BOARD;
     }
